Перламутровые пуговицы, правки
